quick refactor of tags
