+++ scratch.js
@@ -6,7 +6,7 @@ var nlp = require('./src/index');
 // var nlp = require('./builds/helloCompiled');
 // const corpus = require('nlp-corpus');
 // let sotu = corpus.sotu.parsed()[23];
-// const fresh = require('./test/unit/lib/freshPrince.js');
+const fresh = require('./test/unit/lib/freshPrince.js');
 
 // bug.1
 //  .? vs *
@@ -19,21 +19,14 @@ var nlp = require('./src/index');
 // let r = nlp('john and jill were nice');
 // r.debug();
 
-// console.time('parse');
-// let r = nlp(fresh);
-// console.timeEnd('parse');
-//
-// console.time('match');
-// r.match('#Determiner (story|thing|#Adjective)', true);
-// console.timeEnd('match');
-//
-// console.time('tag');
-// r.tag('#Person');
-// console.timeEnd('tag');
-// m.debug()
-//
-// // let m = r.list[0].lookup('story', true);
-// console.log(m.length);
-// console.timeEnd('m');
+console.time('parse');
+let r = nlp(fresh);
+console.timeEnd('parse');
 
-nlp('he is quick').sentences().toFutureTense().debug();
+console.time('match');
+r.match('#Determiner (story|thing|#Adjective)', true);
+console.timeEnd('match');
+
+console.time('tag');
+r.tag('#Person');
+console.timeEnd('tag');
