+++ src/result/methods/loops.js
@@ -21,8 +21,8 @@ let foreach = [
 
   'tagger',
 
-// 'tag',
-// 'unTag',
+  'tag',
+  'unTag',
 ];
 
 const addMethods = (Text) => {
